commit 3

+++ src/main/java/payloads/NewFatturaDTO.java
@@ -4,6 +4,6 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 public record NewFatturaDTO(
-LocalDate data, Long importo, Long numero, Long id
+LocalDate data, Double importo, Long numero, Long id
 ){
 }
